make druidapi pip installable
