+++ examples/quickstart/jupyter-notebooks/druidapi/druidapi/consts.py
